Test Entites moved under test directory.
