+++ test-projects/test-java-project/build.gradle
@@ -2,5 +2,6 @@ group 'com.github.vilmosnagy.elq'
 
 dependencies {
     compileOnly "org.projectlombok:lombok:1.16.10"
+    testCompileOnly "org.projectlombok:lombok:1.16.10"
 }
