+++ test-projects/test-java-project/src/test/java/com/github/vilmosnagy/elq/testproject/entites/Album.java
@@ -1,4 +1,5 @@
-package com.github.vilmosnagy.elq.testproject.entities;
+package com.github.vilmosnagy.elq.testproject.entites;
+
 
 import lombok.Data;
 import lombok.EqualsAndHashCode;
