+++ test-projects/test-kotlin-project/src/test/kotlin/com/github/vilmosnagy/elq/testproject/entities/Album.kt
