+++ test-projects/test-kotlin-project/src/test/kotlin/com/github/vilmosnagy/elq/testproject/entitytests/AlbumTest.kt
@@ -1,10 +1,12 @@
-package com.github.vilmosnagy.elq.testproject.entities
+package com.github.vilmosnagy.elq.testproject.entitytests
 
 import com.avaje.ebean.Ebean
 import com.avaje.ebean.Expr
 import com.github.vilmosnagy.elq.elqcore.interfaces.Predicate
 import com.github.vilmosnagy.elq.elqcore.stream.createElqStream
 import com.github.vilmosnagy.elq.testproject.BaseTest
+import com.github.vilmosnagy.elq.testproject.entities.Album
+import com.github.vilmosnagy.elq.testproject.entities.Artist
 import org.junit.Assert.*
 import java.util.stream.Collectors
